fixed 2 bugs (see details below)
built-in predicate add is now properly weakened. before, Flowlog was attempting to partially evaluate it.

also fixed a separate bug in Partial_Eval's invocation of reassemble_xsb_equality: if we had an atomic formula that looked like P(X, 3), we'd get a map2 exception. This was caused by failing to remove the constant terms from the list before passing to r_x_e.